Many changes. Physics objects Gesture recognizer test Playgo hierarchy Edge Detection

+++ poly2tri/common/shapes.cc
@@ -119,8 +119,8 @@ void Triangle::ClearDelunayEdges()
 Point* Triangle::OppositePoint(Triangle& t, Point& p)
 {
   Point *cw = t.PointCW(p);
-  double x = cw->x;
-  double y = cw->y;
+  float x = cw->x;
+  float y = cw->y;
   x = p.x;
   y = p.y;
   return PointCW(*cw);
