+++ poly2tri/common/shapes.h
@@ -44,7 +44,7 @@ namespace p2t {
 
 	struct Point {
 
-		double x, y;
+		float x, y;
 
 		/// Default constructor does nothing (for performance).
 		Point()
@@ -57,7 +57,7 @@ namespace p2t {
 		std::vector<Edge*> edge_list;
 
 		/// Construct using coordinates.
-		Point(double x, double y) : x(x), y(y) {}
+		Point(float x, float y) : x(x), y(y) {}
 
 		/// Set this point to all zeros.
 		void set_zero()
@@ -67,7 +67,7 @@ namespace p2t {
 		}
 
 		/// Set this point to some specified coordinates.
-		void set(double x_, double y_)
+		void set(float x_, float y_)
 		{
 			x = x_;
 			y = y_;
@@ -96,22 +96,22 @@ namespace p2t {
 		}
 
 		/// Multiply this point by a scalar.
-		void operator *=(double a)
+		void operator *=(float a)
 		{
 			x *= a;
 			y *= a;
 		}
 
 		/// Get the length of this point (the norm).
-		double Length() const
+		float Length() const
 		{
 			return sqrt(x * x + y * y);
 		}
 
 		/// Convert this point into a unit point. Returns the Length.
-		double Normalize()
+		float Normalize()
 		{
-			double len = Length();
+			float len = Length();
 			x /= len;
 			y /= len;
 			return len;
@@ -242,7 +242,7 @@ namespace p2t {
 	}
 
 	/// Multiply point by scalar
-	inline Point operator *(double s, const Point& a)
+	inline Point operator *(float s, const Point& a)
 	{
 		return Point(s * a.x, s * a.y);
 	}
@@ -258,27 +258,27 @@ namespace p2t {
 	}
 
 	/// Peform the dot product on two vectors.
-	inline double Dot(const Point& a, const Point& b)
+	inline float Dot(const Point& a, const Point& b)
 	{
 		return a.x * b.x + a.y * b.y;
 	}
 
 	/// Perform the cross product on two vectors. In 2D this produces a scalar.
-	inline double Cross(const Point& a, const Point& b)
+	inline float Cross(const Point& a, const Point& b)
 	{
 		return a.x * b.y - a.y * b.x;
 	}
 
 	/// Perform the cross product on a point and a scalar. In 2D this produces
 	/// a point.
-	inline Point Cross(const Point& a, double s)
+	inline Point Cross(const Point& a, float s)
 	{
 		return Point(s * a.y, -s * a.x);
 	}
 
 	/// Perform the cross product on a scalar and a point. In 2D this produces
 	/// a point.
-	inline Point Cross(const double s, const Point& a)
+	inline Point Cross(const float s, const Point& a)
 	{
 		return Point(-s * a.y, s * a.x);
 	}
